Changing default rates histogram
I noticed that when I changed to Python3, the labels for the x-axis are not show for the line plots.
I have not been able to figure that out and I had to display the default chart as histogram. In this version I also added a table showing the default rate per bin